feat(surveys): add back button option to the survey editor
Exposes the opt-in back button (posthog-js appearance.allowGoBack) in the
survey editor:

- New shared SurveyBehaviorOptions component (shuffle + allow-go-back), used
  by both the classic editor and the guided wizard's appearance step.
- Optional back button label (backButtonText) with a live editor preview.
- Hosted survey page styling for the back button.
- Submit and back button labels are now translatable (LLM generation,
  editor UI, and preview).

Generated-By: PostHog Code
Task-Id: 9f99f4a8-7f06-4cc2-a01b-20292675c2e8
