+++ frontend/src/scenes/surveys/surveyTranslationUtils.ts
@@ -22,6 +22,12 @@ export function getSurveyWithTranslatedContent<TSurvey extends Survey | NewSurve
     if (translation.thankYouMessageCloseButtonText) {
         appearanceUpdates.thankYouMessageCloseButtonText = translation.thankYouMessageCloseButtonText
     }
+    if (translation.submitButtonText) {
+        appearanceUpdates.submitButtonText = translation.submitButtonText
+    }
+    if (translation.backButtonText) {
+        appearanceUpdates.backButtonText = translation.backButtonText
+    }
 
     return {
         ...survey,
