+++ frontend/src/scenes/surveys/wizard/steps/AppearanceStep.tsx
@@ -27,6 +27,7 @@ import {
     defaultSurveyAppearance,
     getMatchingSurveyThemeId,
 } from '../../constants'
+import { SurveyBehaviorOptions } from '../../survey-appearance/SurveyBehaviorOptions'
 import { SurveyAppearancePreview } from '../../SurveyAppearancePreview'
 import { surveyLogic } from '../../surveyLogic'
 import { surveysLogic } from '../../surveysLogic'
@@ -35,8 +36,8 @@ import { SurveyThemeSelector } from '../SurveyThemeSelector'
 import { WizardSection } from '../WizardLayout'
 
 export function AppearanceStep(): JSX.Element {
-    const { survey } = useValues(surveyLogic)
-    const { setSurveyValue } = useActions(surveyLogic)
+    const { survey, hasBranchingLogic } = useValues(surveyLogic)
+    const { setSurveyValue, deleteBranchingLogic } = useActions(surveyLogic)
     const { surveysStylingAvailable } = useValues(surveysLogic)
     const { guardAvailableFeature } = useValues(upgradeModalLogic)
     const { isDarkModeOn } = useValues(themeLogic)
@@ -170,6 +171,17 @@ export function AppearanceStep(): JSX.Element {
                         }
                     }}
                 />
+
+                {/* Question behavior */}
+                <WizardSection title="Behavior" titleClassName="text-sm font-medium">
+                    <SurveyBehaviorOptions
+                        survey={survey}
+                        onAppearanceChange={onAppearanceChange}
+                        hasBranchingLogic={hasBranchingLogic}
+                        deleteBranchingLogic={deleteBranchingLogic}
+                        onTranslationsChange={(translations) => setSurveyValue('translations', translations)}
+                    />
+                </WizardSection>
 
                 {/* Advanced options */}
                 <LemonCollapse
